requirements.txt
imutils==0.5.3
matplotlib==3.3.2
numpy==1.19.2
tensorflow_gpu==2.3.0
Keras==2.4.3
Pillow==8.0.1
scikit_learn==0.23.2
tensorflow==2.3.1
